+++ packages/nesd/lib/ui/save_states/save_states_screen.dart
@@ -116,6 +116,7 @@ class SaveStatesScreen extends HookConsumerWidget {
         child: NesdMenuWrapper(
           child: FocusChild(
             autofocus: true,
+            wrapAround: true,
             child: PaginatedGrid(
               children: [
                 if (nextOpenSlot < 10 && nesController.isOn)